Remove useless component and merge it in to admin order page
